# openapi.yaml
openapi: 3.0.3
info:
  title: Movie Reservation System
  version: 1.0.0
  description: |
    Every successful response has the shape `{ message, data }`.
    Every error response has the shape `{ error, message? }` — `error` is a stable error code, `message` only appears when extra explanation is needed.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
    Genre:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Sci-Fi }
        deletedAt: { type: string, format: date-time, nullable: true }
    Movie:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string, example: Blade Runner 2049 }
        description: { type: string }
        posterImageId: { type: string, nullable: true }
        posterUrl:
          type: string
          nullable: true
          description: Full image URL derived from posterImageId
          example: http://localhost:4000/api/image/poster-1234567890-123456789.jpg
        genreId: { type: string, format: uuid }
        genre: { $ref: '#/components/schemas/Genre' }
        deletedAt: { type: string, format: date-time, nullable: true }
    Theater:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Cinema 1 }
    Seat:
      type: object
      properties:
        id: { type: string, format: uuid }
        theaterId: { type: string, format: uuid }
        row: { type: string, example: A }
        number: { type: integer, example: 7 }
        type: { type: string, enum: [REGULAR, VIP] }
...
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        totalPages: { type: integer, example: 3 }
    Error:
      type: object
      properties:
        error: { type: string, description: Stable error code, example: GenreNotFound }
        message: { type: string, description: Only present when extra explanation is needed }
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            issues:
              type: array
              description: Zod issue list, only present when error is ValidationError
              items: { type: object }
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
    Unauthorized:
      description: Missing cookie, invalid token, or a token revoked by logout
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Unauthorized }
    Forbidden:
      description: Authenticated but not an ADMIN
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Forbidden }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - cookieAuth: []
paths:
...
  /api/movies:
    get:
      tags: [Movies]
      summary: List movies
      security: []
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: genreId, in: query, schema: { type: string, format: uuid } }
        - { name: search, in: query, description: 'Case-insensitive match on title', schema: { type: string }, example: godfather }
        - { name: orderBy, in: query, description: Ordered by creation time, schema: { type: string, enum: [asc, desc], default: desc } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Movie' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/movies/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Movies]
      summary: Get a movie
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Movie' }
        '404':
          description: MovieNotFound
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: MovieNotFound }
  /api/genres:
    get:
      tags: [Genres]
      summary: List genres
      security: []
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: orderBy, in: query, description: Ordered by name, schema: { type: string, enum: [asc, desc], default: desc } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Genre' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/genres/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Genres]
      summary: Get a genre
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Genre' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/theaters:
    get:
      tags: [Theaters]
      summary: List theaters
      security: []
      parameters:
        - { name: name, in: query, description: 'Case-insensitive match on name', schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 10 } }
        - { name: orderBy, in: query, description: Ordered by name, schema: { type: string, enum: [asc, desc], default: asc } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Theater' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/theaters/{theaterId}:
    parameters:
      - { name: theaterId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Theaters]
      summary: Get a theater
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Theater' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/admin/movies:
    post:
      tags: [Admin]
      summary: Create a movie
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, description, genreId, poster]
              properties:
                title: { type: string, example: Blade Runner 2049 }
                description: { type: string, example: A young blade runner uncovers a secret that could tip the balance of society. }
                genreId: { type: string, format: uuid }
                poster: { type: string, format: binary, description: 'jpeg/png/gif, up to 10MB' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Movie' }
        '400':
          description: '`PosterRequired`, `GenreNotFound` (genre missing or soft-deleted), `UploadError`, or `ValidationError`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: PosterRequired, message: Poster image is required }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/movies/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Admin]
      summary: Update a movie — at least one field or a poster
      description: When the poster changes, the previous file is deleted from disk after a successful update.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title: { type: string, example: "Blade Runner 2049 (Director's Cut)" }
                description: { type: string }
                genreId: { type: string, format: uuid }
                poster: { type: string, format: binary }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Movie' }
        '400':
          description: '`EmptyUpdate` or `GenreNotFound`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: EmptyUpdate, message: At least one field or a poster is required }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Admin]
      summary: Soft-delete a movie
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/admin/genres:
    post:
      tags: [Admin]
      summary: Create a genre
      description: If the name matches a soft-deleted genre, that genre is restored instead of returning an error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: Documentary }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Genre' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: GenreAlreadyExists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: GenreAlreadyExists }
  /api/admin/genres/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Admin]
      summary: Rename a genre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: Docu-drama }
      responses:
        '200': { description: Updated }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: GenreAlreadyExists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Admin]
      summary: Soft-delete a genre
      description: Refused while any non-deleted movie still uses this genre.
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: GenreInUse
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: GenreInUse, message: 3 movie(s) still use this genre }
  /api/admin/theaters:
    post:
      tags: [Admin]
      summary: Create a theater
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: Cinema 1 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Theater' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: TheaterAlreadyExists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: TheaterAlreadyExists }
  /api/admin/theaters/{theaterId}:
    parameters:
      - { name: theaterId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Admin]
      summary: Rename a theater
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: Cinema 2 }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Theater' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: TheaterAlreadyExists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: TheaterAlreadyExists }
  /api/admin/theaters/{theaterId}/layout:
    parameters:
      - { name: theaterId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Admin]
      summary: Replace a theater's seat layout
      description: |
        Reconciles the theater's seats with `rows`, inside a single transaction. Seats are keyed by (row, number),
        so a seat that stays in the layout keeps its id and its reservations — only its `type` is updated.
        Seats missing from the new layout are deleted, and the whole request is refused with `SeatsInUse` if any of them is already reserved.
        Each row produces seats numbered 1..seatCount; a seat inside an `overrides` range takes that override's `type`, otherwise it takes `defaultType`.
        Override ranges must not overlap and must not exceed `seatCount`; row names must be unique.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rows]
              properties:
                rows:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [name, seatCount]
                    properties:
                      name: { type: string, description: Upper-cased automatically, example: A }
                      seatCount: { type: integer, minimum: 1, maximum: 100, example: 10 }
                      defaultType: { type: string, enum: [REGULAR, VIP], default: REGULAR }
                      overrides:
                        type: array
                        default: []
                        items:
                          type: object
                          required: [from, to, type]
                          properties:
                            from: { type: integer, minimum: 1, example: 8 }
                            to: { type: integer, minimum: 1, example: 10 }
                            type: { type: string, enum: [REGULAR, VIP] }
            example:
              rows:
                - { name: A, seatCount: 10, overrides: [{ from: 8, to: 10, type: VIP }] }
                - { name: B, seatCount: 12, defaultType: VIP }
      responses:
        '200':
          description: The seat layout after the replacement
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Seat' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: SeatsInUse — a seat dropped by the new layout already has a reservation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: SeatsInUse }
...